+++ CHANGES.md
@@ -8,6 +8,7 @@ Version 0.8.4 (Pending)
 - Renamed Staff Report to Users Report
 - Added Staff to Check in by Badge report
 - Added checked in count to Staff Check In screen
+- Added utility/testbadges screen
 
 Version 0.8.3 (10/9/2019)
 - Changed "Registration" in menu bar to link to home page